Add files via upload
In this certificate, the User name will be fetched from the database and written on the certificate, and the type of honour, where CODE OF HONOUR, either Student, Leaders and Citizens, for now, the date of the time of certification will also be written, while the signature of the organization will be written. All these written texts will be fetched from the database.